Вернул интерфейсы в Form1
